Reorganize and cleanup compiler code.
