+++ core/compiler/src/main/scala/com/fortysevendeg/exercises/compiler/DocParser.scala
@@ -1,12 +1,13 @@
 package com.fortysevendeg.exercises
 package compiler
 
-import scala.tools.nsc.util.DocStrings
+import scala.reflect.internal.Chars.isWhitespace
 
-import cats._
 import cats.data.Xor
-import cats.std.all._
 
+/** Handles parsing doc comment strings into friendly data structures
+  * containing the relevant information need by the exercise compiler.
+  */
 object DocParser {
 
   case class ParsedLibraryComment(
@@ -77,20 +78,45 @@ object DocParser {
     }
   }
 
-  private def skipToEolEx(str: String, start: Int): Int =
+  // ~ BEGIN
+  // The following methods are strongly based off of code in
+  // scala.tools.nsc.util.DocStrings. I chose to copy the methods here
+  // so adjustments could be made for our purposes. A bastardization, of sorts.
+
+  // Adjusted from DocStrings.skipToEol
+  private def skipToEol(str: String, start: Int): Int =
     if (start + 2 < str.length && (str charAt start) == '/' && (str charAt (start + 1)) == '*' && (str charAt (start + 2)) == '*') start + 3
     else if (start + 1 < str.length && (str charAt start) == '*' && (str charAt (start + 1)) == '/') start
-    else if (start < str.length && (str charAt start) != '\n') skipToEolEx(str, start + 1)
+    else if (start < str.length && (str charAt start) != '\n') skipToEol(str, start + 1)
     else start
 
+  // Our own creation
   private def trimTrailingWhitespace(str: String, start: Int, end: Int): Int =
-    if (end >= start && (str charAt (end - 1)) == ' ') trimTrailingWhitespace(str, start, end - 1)
+    if (end >= start && isWhitespace(str charAt (end - 1))) trimTrailingWhitespace(str, start, end - 1)
     else end
+
+  // Verbatim copy of DocStrings.skipLineLead
+  private def skipLineLead(str: String, start: Int): Int =
+    if (start == str.length) start
+    else {
+      val idx = skipWhitespace(str, start + 1)
+      if (idx < str.length && (str charAt idx) == '*') skipWhitespace(str, idx + 1)
+      else if (idx + 2 < str.length && (str charAt idx) == '/' && (str charAt (idx + 1)) == '*' && (str charAt (idx + 2)) == '*')
+        skipWhitespace(str, idx + 3)
+      else idx
+    }
+
+  // Verbatim copy of DocStrings.skipWhitespace
+  private def skipWhitespace(str: String, start: Int): Int =
+    if (start < str.length && isWhitespace(str charAt start)) skipWhitespace(str, start + 1)
+    else start
+
+  // ~ END
 
   private def cleanLines(lines: List[String]): List[String] = {
     lines.map { line ⇒
-      val ll = DocStrings.skipLineLead(line, -1)
-      val le0 = skipToEolEx(line, ll)
+      val ll = skipLineLead(line, -1)
+      val le0 = skipToEol(line, ll)
       val le1 = trimTrailingWhitespace(line, ll, le0)
       line.substring(ll, le1)
     }
